better mpsdiff

+++ include/calibration.h
@@ -4,7 +4,7 @@
 int calculateMinPartition(int* v, int size, int minCost, double a, double b, double c, int seed);
 void adjustSearchRange(int minPartitionSize, int* minMPS, int* maxMPS, int* steps, int* mps, int numMPS);
 
-int calculateMinBreaks(int* v, int size, float minCost, int minPartitionSize, double a, double b, double c, int seed);
+int calculateMinBreaks(int* v, int size, double minCost, int minPartitionSize, double a, double b, double c, int seed);
 
 
 #endif
